chore: split into containers and tags
