+++ lib/src/parser.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 
 import 'package:audio_metadata_reader/audio_metadata_reader.dart';
-import 'package:audio_metadata_reader/src/parsers/riff.dart';
+import 'package:audio_metadata_reader/src/parsers/containers/riff.dart';
 import 'package:audio_metadata_reader/src/metadata/base.dart';
 
 /// Parse the metadata of a file.
