+++ lib/src/parsers/containers/flac.dart
@@ -3,10 +3,10 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:audio_metadata_reader/src/metadata/base.dart';
-import 'package:audio_metadata_reader/src/parsers/tag_parser.dart';
+import 'package:audio_metadata_reader/src/parsers/tags/tag_parser.dart';
 import 'package:audio_metadata_reader/src/utils/buffer.dart';
 
-import '../utils/bit_manipulator.dart';
+import 'package:audio_metadata_reader/src/utils/bit_manipulator.dart';
 
 /// The different reserved block types that are defined
 /// for this format
