+++ lib/src/parsers/containers/mp4.dart
@@ -3,12 +3,12 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:audio_metadata_reader/src/metadata/base.dart';
-import 'package:audio_metadata_reader/src/parsers/tag_parser.dart';
+import 'package:audio_metadata_reader/src/parsers/tags/tag_parser.dart';
 import 'package:audio_metadata_reader/src/utils/buffer.dart';
 import 'package:mime/mime.dart';
 
-import '../../audio_metadata_reader.dart';
-import '../utils/bit_manipulator.dart';
+import 'package:audio_metadata_reader/audio_metadata_reader.dart';
+import 'package:audio_metadata_reader/src/utils/bit_manipulator.dart';
 
 // https://xhelmboyx.tripod.com/formats/mp4-layout.txt
