+++ lib/src/parsers/containers/ogg.dart
@@ -2,8 +2,8 @@ import 'dart:io';
 import 'dart:typed_data';
 
 import 'package:audio_metadata_reader/src/metadata/base.dart';
-import 'package:audio_metadata_reader/src/parsers/tag_parser.dart';
-import 'package:audio_metadata_reader/src/parsers/vorbis_comment.dart';
+import 'package:audio_metadata_reader/src/parsers/tags/tag_parser.dart';
+import 'package:audio_metadata_reader/src/parsers/tags/vorbis_comment.dart';
 import 'package:audio_metadata_reader/src/utils/bit_manipulator.dart';
 import 'package:audio_metadata_reader/src/utils/buffer.dart';
